refactor(v2): move `reportMessage` from `core/src/server/utils` to `utils` package
